Fixes of fails
